fixed based on pr, added validators for request, check if user is already in the room - used in joining/leaving rooms

+++ wt_mobile/backend_logic_rooms.py
@@ -4,6 +4,7 @@ from .models import Room, User, UserRoom
 from .serializers import RoomSerializer, JoinedRoomSerializer
 from django.core.exceptions import MultipleObjectsReturned
 from watch_together.general_utils import get_loggers
+from uuid import UUID
 
 # ---------Defines--------- #
 
@@ -26,6 +27,10 @@ def create_room(request) -> dict:
     name = request.data.get('name')
     owner = request.data.get('owner_id')
     password = request.data.get('password')
+
+    # Check request data validity
+    if not is_uuid(unique_id) or not is_int(owner):
+        return ERROR_MSG
 
     # Check if room owner is existing user in the database (should be existing)
     if not User.objects.filter(id=owner).exists():
@@ -63,12 +68,16 @@ def delete_room(request) -> dict:
     # Get room unique ID
     unique_id = request.data.get('unique_id')
 
+    # Check request data validity
+    if not is_uuid(unique_id):
+        return ERROR_MSG
+
     # Try deleting room
     try:
         Room.objects.get(unique_id=unique_id).delete()
-        return {SUCCESS: 'Room deleted'}
     except (Room.DoesNotExist, MultipleObjectsReturned):
         return ERROR_MSG
+    return {SUCCESS: 'Room deleted'}
 
 
 def edit_room(request) -> dict:
@@ -79,6 +88,10 @@ def edit_room(request) -> dict:
     new_name = request.data.get('new_name')
     new_password = request.data.get('new_password')
 
+    # Check request data validity
+    if not is_uuid(unique_id):
+        return ERROR_MSG
+
     try:
         if Room.objects.filter(unique_id=unique_id).exists():
             room_to_edit = Room.objects.get(unique_id=unique_id)
@@ -87,9 +100,9 @@ def edit_room(request) -> dict:
             if new_password:
                 room_to_edit.password = make_password(new_password)
             room_to_edit.save()
-            return {SUCCESS: 'Room edited'}
     except (Room.DoesNotExist, MultipleObjectsReturned):
         return ERROR_MSG
+    return {SUCCESS: 'Room edited'}
 
 
 def get_room() -> dict:
@@ -98,18 +111,23 @@ def get_room() -> dict:
     try:
         all_rooms_obj = Room.objects.all()
         serialized = RoomSerializer(all_rooms_obj, many=True)
-        return serialized
     except Room.DoesNotExist:
         return {ERROR: 'Room does not exist'}
+    return serialized
 
 
 # ---------JoinRoom--------- #
 
 def join_room(request) -> dict:
     """Join a room"""
+
     # Get request data
     user_to_join = request.data.get('user')
     room_to_join = request.data.get('room')
+    # Check if request 'user' is ID and 'room' is UUID
+    if not isinstance(user_to_join, int) or not is_uuid(room_to_join):
+        return ERROR_MSG
+    
     password_input = request.data.get('password')
 
     # Check if request data exists
@@ -118,6 +136,10 @@ def join_room(request) -> dict:
         room = Room.objects.get(unique_id=room_to_join)
     except (User.DoesNotExist, Room.DoesNotExist, MultipleObjectsReturned):
         return ERROR_MSG
+
+    # Check if the user is already joined to that room
+    if UserRoom.objects.filter(room_id=room_to_join, user_id=user_to_join).exists():
+        return {ERROR: "User already in the room"}
 
     # Check if password from request == room password, if there is one
     is_password_matching = False
@@ -134,9 +156,9 @@ def join_room(request) -> dict:
     if is_password_matching:
         # TODO: Check if the user already in the party
         room.users.add(user)
-        return {SUCCESS: 'User joined the room'}
     else:
         return {ERROR: 'Provided password did not match the room password'}
+    return {SUCCESS: 'User joined the room'}
 
 
 def leave_room(request) -> dict:
@@ -145,21 +167,29 @@ def leave_room(request) -> dict:
     user_to_leave = request.data.get('user')
     room_to_leave = request.data.get('room')
 
+    # Check request data validity
+    if not is_uuid(room_to_leave) or not is_int(user_to_leave):
+        return ERROR_MSG
+
     # Get room and user from db
     try:
         user = User.objects.get(id=user_to_leave)
         room = Room.objects.get(unique_id=room_to_leave)
 
+        # Check if the user is in that room
+        if not UserRoom.objects.filter(room_id=room_to_leave, user_id=user_to_leave).exists():
+            return {ERROR: "User is not in the room"}
+
         # Remove the user from the room
         room.users.remove(user)
 
         # If the room is empty, it will not exist in UserRoom db. Delete the room if so
-        if not UserRoom.objects.filter(unique_id=room.unique_id).exists():
+        if not UserRoom.objects.filter(room_id=room.unique_id).exists():
             room.delete()
             dev_logger.info('Room deleted, because all users left')
-            return {SUCCESS: 'User left the room'}
     except (User.DoesNotExist, Room.DoesNotExist, MultipleObjectsReturned):
         return ERROR_MSG
+    return {SUCCESS: 'User left the room'}
 
 
 def list_rooms_user_participates(request) -> dict:
@@ -167,12 +197,37 @@ def list_rooms_user_participates(request) -> dict:
 
     user_id = request.data.get('user')
 
+    # Check if user ID is an integer
+    if not is_int(user_id):
+        return ERROR_MSG
+
     # Return all rooms, that user participates in, if the user exists
     try:
         user = User.objects.get(id=user_id)
         all_rooms_for_user = UserRoom.objects.filter(user_id=user.pk)
         serialized = JoinedRoomSerializer(all_rooms_for_user, many=True)
 
-        return serialized
     except (User.DoesNotExist, Room.DoesNotExist, MultipleObjectsReturned):
         return ERROR_MSG
+    return serialized
+
+
+# ---------Support Functions--------- #
+
+def is_uuid(id):
+    """Check if data=uuid"""
+    try:
+        UUID(str(id))
+
+        return True
+    except ValueError:
+        return False
+
+
+def is_int(id):
+    """Check if provided data is int, because pk are integers"""
+
+    if isinstance(id, int):
+        return True
+    else:
+        return False
